+++ docs/HISTORY.rst
@@ -6,6 +6,7 @@ Changelog
 0.8.1 - unreleased
 ------------------
 
+- https://github.com/collective/plone.jsonapi.routes/issues/48: Implement cut/copy/paste routes
 - https://github.com/collective/plone.jsonapi.routes/issues/46: Route Provider `portal` throws TypeError
 - https://github.com/collective/plone.jsonapi.routes/issues/47: ZCML directive to enable AdvancedQuery if installed